build fix
svn:sync-xref-src-repo-rev: r59733

+++ src/VBox/Runtime/common/checksum/ipv6.cpp
@@ -1,4 +1,4 @@
-/* $Id: ipv6.cpp 28024 2010-04-07 06:34:16Z knut.osmundsen@oracle.com $ */
+/* $Id: ipv6.cpp 28026 2010-04-07 06:43:54Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - IPv6 Checksum calculation and validation.
  */
@@ -43,7 +43,7 @@
  * @copydoc RTNetIPv6PseudoChecksumBits
  */
 DECLINLINE(uint32_t) rtNetIPv6PseudoChecksumBits(PCRTNETADDRIPV6 pSrcAddr, PCRTNETADDRIPV6 pDstAddr,
-                                                 uint8_t bProtocol, uint16_t cbPkt)
+                                                 uint8_t bProtocol, uint32_t cbPkt)
 {
     uint32_t u32Sum = pSrcAddr->au16[0]
                     + pSrcAddr->au16[1]
